backend/requirements.txt
fastapi
uvicorn[standard]
websockets==13.1
python-dotenv
deepgram-sdk==3.1.2
groq
elevenlabs
sentence-transformers
faiss-cpu
numpy